requirements.txt
numpy
pyopencl
pillow
rpyc
